requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-groq>=0.1.0
langchain-cohere>=0.1.0
pdfplumber>=0.10.3
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf2>=3.0.1

